src/backend/requirements.txt
# DA-MACAÉ v2 Backend Dependencies
# Python 3.12+

# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-multipart>=0.0.18

# Azure AI Agent Framework
azure-ai-projects>=1.0.0b7
azure-ai-agents>=1.0.0b3
azure-identity>=1.19.0
aiohttp>=3.10.0

# Azure OpenAI
openai>=1.58.0

# Anthropic (Claude)
anthropic>=0.49.0

# Azure Cosmos DB
azure-cosmos>=4.9.0

# Azure Storage
azure-storage-blob>=12.24.0

# Azure AI Search
azure-search-documents>=11.6.0

# Pydantic
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0

# WebSocket
websockets>=14.0

# MCP
mcp>=1.1.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0
structlog>=24.4.0
tenacity>=9.0.0

# Observability
azure-monitor-opentelemetry>=1.6.4
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
httpx>=0.28.0
